+++ codegen/src/test/kotlin/software/amazon/smithy/rust/codegen/smithy/protocols/parse/EventStreamUnmarshallerGeneratorTest.kt
@@ -14,6 +14,7 @@ import software.amazon.smithy.rust.codegen.smithy.protocols.EventStreamTestModel
 import software.amazon.smithy.rust.codegen.smithy.protocols.EventStreamTestTools
 import software.amazon.smithy.rust.codegen.testutil.TestRuntimeConfig
 import software.amazon.smithy.rust.codegen.testutil.compileAndTest
+import software.amazon.smithy.rust.codegen.testutil.testRustSettings
 import software.amazon.smithy.rust.codegen.testutil.unitTest
 
 class EventStreamUnmarshallerGeneratorTest {
@@ -28,7 +29,8 @@ class EventStreamUnmarshallerGeneratorTest {
             TestRuntimeConfig,
             test.serviceShape,
             ShapeId.from(testCase.protocolShapeId),
-            "test"
+            "test",
+            testRustSettings(test.model)
         )
         val protocol = testCase.protocolBuilder(codegenContext)
         val generator = EventStreamUnmarshallerGenerator(