+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/CodegenVisitor.kt
@@ -64,8 +64,7 @@ class CodegenVisitor(context: PluginContext, private val codegenDecorator: RustC
         val baseProvider = RustCodegenPlugin.baseSymbolProvider(model, service, symbolVisitorConfig)
         symbolProvider = codegenDecorator.symbolProvider(generator.symbolProvider(model, baseProvider))
 
-        codegenContext =
-            CodegenContext(model, symbolProvider, settings.runtimeConfig, service, protocol, settings.moduleName)
+        codegenContext = CodegenContext(model, symbolProvider, service, protocol, settings)
         rustCrate = RustCrate(
             context.fileManifest,
             symbolProvider,