+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 vNext (Month Day, Year)
 =======================
 
-**New this week**
+- Add `RustSettings` to `CodegenContext` (smithy-rs#616, smithy-rs#752)
 
+v0.25.1 (October 11th, 2021)
+=========================
+**New this week**
 - :bug: Re-add missing deserialization operations that were missing because of a typo in `HttpBoundProtocolGenerator.kt`
 
 v0.25 (October 7th, 2021)